Update starter kits and data loading utilities.
Updated starter kit so labels for challenge points are not passed through the ChallengeDataset because they are not available for dev and eval sets. Instead defined function that takes in train_data_path and returns attack model that will assign confidence scores to challenge points. Updated corresponding data utilities in library